MCMC Algorithm

This function executes the Markov Chain Monte Carlo (MCMC) algorithm for STDGLM.

Usage

stdglm(
  y,
  X,
  Z = NULL,
  offset = NULL,
  point.referenced = TRUE,
  random.walk = FALSE,
  W,
  W_pred = NULL,
  W_cross = NULL,
  X_pred = NULL,
  Z_pred = NULL,
  offset_pred = NULL,
  nrep = 100,
  nburn = 100,
  thin = 1,
  print.interval = 100,
  prior = NULL,
  keepY = TRUE,
  keepLogLik = TRUE,
  last_run = NULL
)

Arguments

y

Either a p-by-t matrix corresponding to the response variable, where p is the number of spatial locations and t is the number of time points.

X

A p-by-t-by-ncovx array corresponding to the covariates whose effects should vary across space and time.

Z

A p-by-t-by-ncovz array corresponding to the covariates whose effects are constant across space and time. Defaults to NULL.

offset

A p-by-t matrix corresponding to the offset term. If NULL, it is set to zero.

point.referenced

A logical indicating whether the data are point-referenced (TRUE) or areal (FALSE). Default is TRUE. If FALSE, predictions are not performed.

random.walk

A logical indicating whether the temporal dynamic should be modeled as a random walk (TRUE) or a first-order autoregressive process (FALSE). Default is FALSE.

W

A p-by-p matrix corresponding to the spatial weights matrix. If point.referenced is TRUE, the distance matrix among the observed locations should be provided. If point.referenced is FALSE, the 0/1 adjacency matrix should be provided.

W_pred

A p-by-p matrix corresponding to the spatial weights matrix for the prediction locations. If NULL, predictions are not performed.

W_cross

A p-by-p_new matrix corresponding to the cross spatial weights matrix between the observed and prediction locations.

X_pred

A p_new-by-t_new-by-ncovx array corresponding to the covariates with varying coefficients for predictions, where p_new is the number of prediction locations and t_new=t+h_ahead, h_ahead>=0, is the number of time points for which predictions are to be made.

Z_pred

A p_new-by-t_new-by-ncovz array corresponding to the covariates with constant coefficients for predictions. Defaults to NULL.

offset_pred

A p_new-by-t_new matrix corresponding to the offset term for predictions. If NULL, it is set to zero.

nrep

An integer indicating the number of iterations to keep after burn-in.

nburn

An integer indicating the number of iterations to discard as burn-in.

thin

An integer indicating the thinning value. Default is 1.

print.interval

An integer indicating the interval at which to print progress messages.

prior

A named list containing the hyperparameters of the model. If NULL, default non-informative hyperparameters are used. See details below.

keepY

A logical indicating whether to keep the response variable in the output. Default is TRUE.

keepLogLik

A logical indicating whether to keep the log-likelihood in the output. Default is TRUE.

last_run

An optional list containing the output from a previous run of the function, which can be used to restore the state of the sampler. Default is NULL.

Value

A named list containing out and ave. ave stores posterior summaries, while out stores the MCMC samples.

Details

prior must be a named list with these elements:

V_beta_0

A scalar defining the prior variance for the initial state of the time-varying coefficients, related to the covariates in X.

V_gamma

A scalar defining the prior variance for the initial state of the constant coefficients, related to the covariates in Z. Assign a value even if Z is NULL.

a1

A scalar defining the inverse-gamma prior shape for the temporal variance of the time-varying coefficients.

b1

A scalar defining the inverse-gamma prior rate for the temporal variance of the time-varying coefficients.

s2_a

A scalar defining the inverse-gamma prior shape for the measurement error variance.

s2_b

A scalar defining the inverse-gamma prior rate for the measurement error variance.

References

insert here.